Add Cloudflare Workers edge entry
Amp-Thread-ID: https://ampcode.com/threads/T-019ecfe0-1aa3-74ee-a0dd-f3ed4b7874f2
