+++ package.json
@@ -3,6 +3,18 @@
   "version": "1.1.0-jq-1.8.1",
   "description": "A high-performance jq wrapper using Emscripten and WebAssembly.",
   "main": "dist/index.js",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "require": "./dist/index.js",
+      "default": "./dist/index.js"
+    },
+    "./edge": {
+      "types": "./dist/edge.d.ts",
+      "require": "./dist/edge.js",
+      "default": "./dist/edge.js"
+    }
+  },
   "repository": "github:owenthereal/jq-wasm",
   "homepage": "https://github.com/owenthereal/jq-wasm#readme",
   "bugs": {
