+++ src/wasm.d.ts
@@ -1,0 +1,4 @@
+declare module "*.wasm" {
+  const wasmModule: WebAssembly.Module;
+  export default wasmModule;
+}
